AER-856 manure storage  (#156)
* IMAER changes for new manure storage emission source
Also snuck in some changes for farmland use with regards to spreading manure.
* Tonnes from integer to double
* Added new optional properties for farmland
Updated the roundtrip test to also include cases where these are defined.
Also introduced FarmSourceEmissionsUtil which can help to calculate emissions (should avoid some code duplication when we add manure storage).
* Domain classes for manure storage
* GML base classes for ManureStorage
* Implemented some visit methods
Should handle the validations and emission calculations with regards to manure storage.
* GML specific version code and roundtriptest
* Added tests for emissions calculator and validator for manure storage
* Changed new methods in interfaces to default
Should prevent the need for the current PRs in other projects to have to rebase for a while.

+++ source/imaer-shared/src/main/java/nl/overheid/aerius/shared/emissions/IsFarmEmissionFactorTypeObject.java
@@ -46,11 +46,11 @@ public interface IsFarmEmissionFactorTypeObject {
   default void setMetersCubed(final Double metersCubed) {
   }
 
-  default Integer getTonnes() {
+  default Double getTonnes() {
     return null;
   }
 
-  default void setTonnes(final Integer tonnes) {
+  default void setTonnes(final Double tonnes) {
   }
 
   default Double getMetersSquared() {
